feat: Add certificate for protonmail-bridge
this is done by the internal cluster CA. It certifies the IP of the
loopback device the bridge expects AND the address (or IP) used by the
node the container runs on. All goes to plan this should fix the cert
woes.

Will need to be copied by hand into the bridge. Not a clean solution but
_a_ solution.